+++ cmd/aux.go
@@ -284,3 +284,70 @@ func PathSHA256(root string) ([]ScannedFile, error) {
 	})
 	return files, err
 }
+
+// func DownloadJSONFile(src, dst string) error {
+// 	// Send a HEAD request to get the Content-Type without downloading the body
+// 	resp, err := http.Head(src)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	defer resp.Body.Close()
+
+// 	// Check if the Content-Type is JSON
+// 	// if resp.Header.Get("Content-Type") != "application/json" {
+// 	// 	return errors.New("source is not a JSON file")
+// 	// }
+
+// 	// Configure the getter client
+// 	client := &getter.Client{
+// 		Src:  src,
+// 		Dst:  dst,
+// 		Mode: getter.ClientModeFile,
+// 	}
+
+// 	// Configure the yacspin spinner
+// 	cfg := yacspin.Config{
+// 		Frequency:       100 * time.Millisecond,
+// 		CharSet:         yacspin.CharSets[3],
+// 		Suffix:          " Downloading",
+// 		SuffixAutoColon: true,
+// 		StopCharacter:   "│ ✓",
+// 		StopColors:      []string{"fgGreen"},
+// 	}
+
+// 	spinner, err := yacspin.New(cfg)
+// 	if err != nil {
+// 		return err
+// 	}
+
+// 	// Start the spinner
+// 	err = spinner.Start()
+// 	if err != nil {
+// 		return err
+// 	}
+
+// 	// Download the file
+// 	err = client.Get()
+// 	if err != nil {
+// 		spinner.StopFail()
+// 		return err
+// 	}
+
+// 	// Stop the spinner
+// 	spinner.Stop()
+
+// 	// Read the downloaded file
+// 	data, err := ioutil.ReadFile(dst)
+// 	if err != nil {
+// 		return err
+// 	}
+
+// 	// Compute the SHA-256 hash of the file
+// 	hash := sha256hash(data)
+
+// 	fmt.Println("│")
+// 	fmt.Println("├ Hash: ", hash)
+// 	fmt.Println("│")
+
+// 	return nil
+// }
